Create signup.php
signup.php is a PHP script that allows users to create an account on a website. It takes the following inputs from the user:

Name
Email
Password
Repeat password
The script validates the inputs and ensures that the password and repeat password match. If all of the inputs are valid, the script creates a new user account in the database. If any of the inputs are invalid, the script displays an error message to the user.

The script also displays a link to the login page, so that users can log in to their new accounts.

Here is a more detailed explanation of each section of the script:

PHP
<?php
if (isset($_GET['error'])) {
  if ($_GET['error'] == 'emptyinputs') {
    echo '<p>Fill in all fields!</p>';
  } else if ($_GET['error'] == 'invalidname') {
    echo '<p>Choose a proper username!</p>';
  } else if ($_GET['error'] == 'invalidemail') {
    echo '<p>Choose a proper email!</p>';
  } else if ($_GET['error'] == 'passwordmatch') {
    echo '<p>Passwords do not match!</p>';
  } else if ($_GET['error'] == 'stmtfailed') {
    echo '<p>Something went wrong, please try again!</p>';
  } else if ($_GET['error'] == 'none') {
    echo '<p>You have successfully signed up!</p>';
  }
}
if (isset($_GET['signup'])) {
  if ($_GET['signup'] == 'success') {
    echo '<p>You have successfully signed up!</p>';
  }
}
?>
Use code with caution. Learn more
This section of the script checks if the error or signup GET parameters are set. If the error parameter is set, the script displays the corresponding error message to the user. If the signup parameter is set, the script displays a success message to the user.

PHP
<form action="includes/signup.inc.php" method="post">
  <input type="text" name='name' placeholder="Enter username"><br>
  <input type="email" name='email' placeholder="Enter email"><br>
  <input type="password" name='pwd' placeholder="Enter password"><br>
  <input type="password" name='pwdRepeat' placeholder="Enter password"><br>
  <input type="submit" name='submit' value="Register">
</form>
Use code with caution. Learn more
This section of the script contains a form that allows users to enter their name, email, password, and repeat password. The form submits the data to the includes/signup.inc.php script when the user clicks the "Register" button.

PHP
<a href="login.php">LOGIN</a> <br>
<a href="index.php">HOME</a>

This section of the script contains links to the login page and the home page.